Endpoint deletar usuário

+++ src/services/user/delete-user-service.js
@@ -1,18 +1,20 @@
 import { PrismaClient } from '@prisma/client'
 const prisma = new PrismaClient()
 
-const serviceDeletar = async (id) => {
-    const deletarUser = await prisma.usuario.delete({
-        where: {
-            id
-        }
-    })
+import { appError } from '../../errors/app-errors.js'
 
-    if (deletarUser) {
+const serviceDeletar = async (userId, paramsId) => {
 
+    if (userId != paramsId) {
+        appError("Você não tem permissão para deletar esse usuário.", 403)
+    } else {
+        const deletarUser = await prisma.usuario.delete({
+            where: {
+                id: userId
+            }
+        })
+        return deletarUser
     }
-
-    return deletarUser
 }
 
 export default serviceDeletar
